Add files via upload
XYZ Company has three departments- Assembly,
Painting, and Packing. The company can make three
types of almirahs. An almirah of type I requires one hour
of assembly, 40 minutes of painting, and 20 minutes of
packing time, respectively. Similarly, an almirah of type Il
needs 80 minutes, 20 minutes, and one hour,
respectively. The almirah of type Ill requires 40 minutes
each of assembly, painting and packing time. The total
time available at assembly, painting, and packing
departments is 600 hours, 400 hours, and 800 hours,
respectively. The objective of the XYZ company is to
maximise the profit. The unit profit for types I, Il and Ill is
R$ 40, 80 and 60, respectively.
Suppose that the manager of this XYZ Company is
thinking of renting the production capacities of the three
departments to another almirah manufacturer - ABC
Company. AB¢ Company is interested in minimizing the
rental charges. On the other hand, the XYZ Company
would like to know the worth of production hours to
them, in each of the departments, in order to determine
the rental rates.
(a) Formulated this problem as an LP problem and solved it
using an excel solver to determine the number of each
type of product that should be produced by the XYZ
Company in order to maximize its profit. (b) Formulate
the dual of the primal problem and interpret your results.